Added @Producer, removed some of the duplicate interceptors. Played with some math so the answer to life, the universe and everything is still 42.
