+++ interceptors/interceptors-interim/src/main/java/org/supertribe/interceptors/interim/stack/ClassLevelInterceptorOne.java
