+++ interceptors/interceptors-interim/src/main/java/org/supertribe/interceptors/interim/stack/FullyInterceptedSuperClass.java
@@ -16,16 +16,11 @@
  */
 package org.supertribe.interceptors.interim.stack;
 
-import javax.interceptor.AroundInvoke;
-import javax.interceptor.InvocationContext;
+import javax.interceptor.Interceptors;
 
 /**
  * @version $Rev$ $Date$
  */
-public class MethodLevelInterceptorTwo {
+public class FullyInterceptedSuperClass {
 
-    @AroundInvoke
-    public Object businessMethodInterceptor(final InvocationContext ic) throws Exception {
-        return Utils.addClassSimpleName(ic, this.getClass().getSimpleName());
-    }
 }
