+++ interceptors/interceptors-interim/src/main/java/org/supertribe/interceptors/interim/stack/InterceptorStack.java
@@ -31,6 +31,7 @@ public class InterceptorStack {
     private List<Interception> interceptions = new ArrayList<>();
 
     public InterceptorStack() {
+        System.out.println("Here");
     }
 
     public InterceptorStack add(Interception interception) {
