+++ interceptors/interceptors-interim/src/main/java/org/supertribe/interceptors/interim/stack/SuperClassOfClassLevelInterceptor.java
